refact: helper function to get descriptor from public key
